doc/network: Modify Documentation in the Network group
