+++ firmware/network/net_tools/doc.txt
@@ -1,0 +1,39 @@
+/**
+
+@defgroup network_net_tools Network Tools
+@ingroup  network
+
+### Net_tools
+
+Contains all about networks tools that could be used in others modules,
+this will be a general settings options as the settings of ipv6 address
+as the gets an specified interface (wired or wireless).
+
+The main propose of the network tools is bring to an user manage all network
+options to get the wished configuration. So the most protocols could be require
+do changes in the ipv6 address or needs to manually call a specified interface
+to manage physical configurations.
+
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.c}
+int8_t index;
+index = get_wired_iface(void);
+printf("The interface index is : %d", index); // should be any number if it's -1
+the interface doesn't exist.
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+### Getting an specified interface
+currently the module only could check the wired interface, this should be ETHOS, SLIP or USB_CDC_ECM,
+this function will return the index of the network interface.
+
+### Ipv6 settings
+
+If you want to set an ipv6 address global o multicast, you could use:
+
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~{.c}
+set_ipv6_global(iface_index, ip, prefix); // To global address
+/* or */
+set_ipv6_multicast(iface_index,ip, prefix); // To multicast address
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+
+*/
