+++ firmware/network/udp_server/doc.txt
@@ -2,9 +2,18 @@
 @defgroup udp_server Udp server
 @ingroup  network
 
-### What goes here?
+### Udp server
+
+The Udp server is a operative device (master), that process all request sent by an udp client device. an udp server
+establish a communications within a socket, that requires specify the num of port used by both devices that are communicating.
+The udp server needs to run as a thread, because this process to attend all request generated by different devices, should be repeated
+every time until the server is stopped.
+
+A udp server device could contain an long data storage that could be sent to all clients or receive data from each client connected
+by the socket. Every device using this module works with Ipv6 protocol internet layer.
 
 This module is used to set an udp server. The function ``` udp_server ``` is used to init the server,
 containing the callback function to handle requests.
 
+@image html diagram_udp_server.drawio.png Udp Server, Getting a request and then sends a response width=70%
 */
